% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipAndAddHorizontalDivider.R
\name{clipAndAddHorizontalDivider}
\alias{clipAndAddHorizontalDivider}
\title{Add a horizontal divider}
\usage{
clipAndAddHorizontalDivider(x, dot, n)
}
\arguments{
\item{x}{Data.frame to be modified.}

\item{dot}{Character, used in the divider.}

\item{n}{Integer, how many rows to print with \code{head} and
\code{tail}.}
}
\value{
A data.frame with added horizontal divider, provided the
data.frame has more than \code{n} rows.
}
\description{
For data.frames that have more rows than limited by n, adds
a horizontal divider by replacing the middle row with dots.
}
