% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxColumnWidth.R
\name{maxColumnWidth}
\alias{maxColumnWidth}
\title{Return max column width}
\usage{
maxColumnWidth(x)
}
\arguments{
\item{x}{A data.frame to be processed.}
}
\value{
A vector of max column widths (either value or column
name).
}
\description{
Depending on the class of the column, max of column name or
column content width is returned. If numeric, number of
printing digits is taken into account.
}
