% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareRownames.R
\name{prepareRownames}
\alias{prepareRownames}
\title{Prepare row names}
\usage{
prepareRownames(x)
}
\arguments{
\item{x}{Data.frame for which rownames will be modified}
}
\value{
A data.frame with rows modified (1 becomes "01:" in
a data.frame with more than 9 rows).
}
\description{
Convert rownames so that they are padded 0 to max rowname width
and append colon at the end. Rows 1:10 should now look like 01:
... 10:.
}
