% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peep.R
\name{peep}
\alias{peep}
\title{Fit data.frame print into console width}
\usage{
peep(x, n = 6, digits = 4)
}
\arguments{
\item{x}{Data.frame to be printed.}

\item{n}{Integer. Max number of first and last rows to print.}

\item{digits}{Integer, to what number of digits numeric columns
are being rounded.}
}
\value{
Invisibly returns a data.frame in case you would need it
for some reason. The main side effect is printing of the abridged
version to the console.
}
\description{
Fit data.frame print into console width
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
